reduced warnings

+++ frontend/main.c
@@ -20,7 +20,7 @@
  * Boston, MA 02111-1307, USA.
  */
 
-/* $Id: main.c,v 1.84 2003/11/20 11:16:29 bouvigne Exp $ */
+/* $Id: main.c,v 1.85 2003/12/21 18:41:33 bouvigne Exp $ */
 
 #ifdef HAVE_CONFIG_H
 # include <config.h>
@@ -304,7 +304,7 @@ lame_decoder(lame_global_flags * gfp, FILE * outf, int skip, char *inPath,
 
     if ( 0 == disable_wav_header )
         if (!fseek(outf, 0l, SEEK_SET)) /* if outf is seekable, rewind and adjust length */
-            WriteWaveHeader(outf, wavsize, lame_get_in_samplerate( gfp ),
+            WriteWaveHeader(outf, (int)wavsize, lame_get_in_samplerate( gfp ),
                             tmp_num_channels, 16);
     fclose(outf);
